refactor: admin 가정예배 공지 수정
