+++ apps/admin/src/pages/homeworship/create/index.tsx
@@ -1,0 +1,12 @@
+import HomeworshipCreate from "@/components/homeworship/create";
+import Layout from "@/components/layout";
+
+const HomeWorshipCreatePage = () => {
+  return (
+    <Layout title="가정예배 공지 추가">
+      <HomeworshipCreate />
+    </Layout>
+  );
+};
+
+export default HomeWorshipCreatePage;
